fix: context7 trustScore 타입 float64 (API 가 9.4 같은 소수 반환)

+++ backend/internal/infrastructure/context7/client.go
@@ -28,7 +28,7 @@ type SearchResult struct {
 	ID             string  `json:"id"`
 	Title          string  `json:"title"`
 	Description    string  `json:"description"`
-	TrustScore     int     `json:"trustScore"`
+	TrustScore     float64 `json:"trustScore"`
 	BenchmarkScore float64 `json:"benchmarkScore"`
 	Stars          int     `json:"stars"`
 	TotalSnippets  int     `json:"totalSnippets"`
